+++ src/crds/_utils/startWatchers.ts
@@ -1,33 +1,37 @@
-import { startWatching as startWatchingManagedRealms } from "./realms/handlers.ts";
-import { startWatching as startWatchingClientCredentials } from "./client-credentials/handlers.ts";
+import { startWatching as startWatchingManagedRealms } from "../realms/handlers.ts";
+import { startWatching as startWatchingClientCredentials } from "../client-credentials/handlers.ts";
 
 import {
   scheduleJobNow as scheduleRealmsCleanupJobNow,
   scheduleJobs as scheduleRealmsCleanupJobs,
-} from "./realms/cleanupQueue.ts";
+} from "../realms/cleanupQueue.ts";
 import {
   scheduleJobNow as _scheduleRealmsReconciliationNow,
   scheduleJobs as scheduleRealmsReconciliation,
-} from "./realms/reconciliationQueue.ts";
+} from "../realms/reconciliationQueue.ts";
 
 import {
   scheduleJobNow as scheduleSecretCleanupJobNow,
   scheduleJobs as scheduleSecretCleanupJobs,
-} from "./client-credentials/secretsCleanupQueue.ts";
+} from "../client-credentials/secretsCleanupQueue.ts";
 import {
   scheduleJobNow as _scheduleClientSecretsReconciliationNow,
   scheduleJobs as scheduleClientSecretsReconciliation,
-} from "./client-credentials/reconciliationQueue.ts";
+} from "../client-credentials/reconciliationQueue.ts";
 
-import { genericCrds } from "./genericCrds/genericCrds.ts";
-import { clientsCr } from "./clients/clientsCrd.ts";
+import { clientScopesCr } from "../client-scopes.cr.ts";
+import { clientsCr } from "../clients.cr.ts";
+import { groupsCr } from "../groups.cr.ts";
+import { usersCr } from "../users.cr.ts";
 
 export const startAllWatchers = async () => {
   await Promise.all([
     startWatchingManagedRealms(),
     startWatchingClientCredentials(),
     clientsCr.startWatching(),
-    ...genericCrds.map((c) => c.startWatching()),
+    clientScopesCr.startWatching(),
+    groupsCr.startWatching(),
+    usersCr.startWatching(),
   ]);
 };
 
@@ -42,10 +46,14 @@ export const startAllQueues = async () => {
     clientsCr.scheduleCleanupJobs(),
     clientsCr.scheduleReconciliationJobs(),
 
-    ...genericCrds.map((c) => [
-      c.scheduleCleanupJobs(),
-      c.scheduleReconciliationJobs(),
-    ]).flat(),
+    clientScopesCr.scheduleCleanupJobs(),
+    clientScopesCr.scheduleReconciliationJobs(),
+
+    groupsCr.scheduleCleanupJobs(),
+    groupsCr.scheduleReconciliationJobs(),
+
+    usersCr.scheduleCleanupJobs(),
+    usersCr.scheduleReconciliationJobs(),
   ]);
 
   // Run the cleanups immediatly
@@ -53,6 +61,8 @@ export const startAllQueues = async () => {
     scheduleSecretCleanupJobNow(),
     scheduleRealmsCleanupJobNow(),
     clientsCr.scheduleCleanupJobNow(),
-    ...genericCrds.map((c) => c.scheduleCleanupJobNow()),
+    clientScopesCr.scheduleCleanupJobNow(),
+    groupsCr.scheduleCleanupJobNow(),
+    usersCr.scheduleCleanupJobNow(),
   ]);
 };
